+++ debug_toolbar/panels/cache.py
@@ -218,20 +218,26 @@ class CachePanel(Panel):
     @property
     def nav_subtitle(self):
         cache_calls = len(self.calls)
-        return ungettext(
-            "%(cache_calls)d call in %(time).2fms",
-            "%(cache_calls)d calls in %(time).2fms",
-            cache_calls,
-        ) % {"cache_calls": cache_calls, "time": self.total_time}
+        return (
+            ungettext(
+                "%(cache_calls)d call in %(time).2fms",
+                "%(cache_calls)d calls in %(time).2fms",
+                cache_calls,
+            )
+            % {"cache_calls": cache_calls, "time": self.total_time}
+        )
 
     @property
     def title(self):
         count = len(getattr(settings, "CACHES", ["default"]))
-        return ungettext(
-            "Cache calls from %(count)d backend",
-            "Cache calls from %(count)d backends",
-            count,
-        ) % dict(count=count)
+        return (
+            ungettext(
+                "Cache calls from %(count)d backend",
+                "Cache calls from %(count)d backends",
+                count,
+            )
+            % dict(count=count)
+        )
 
     def enable_instrumentation(self):
         if isinstance(middleware_cache.caches, CacheHandlerPatch):